chore: rename static base controller
