+++ packages/shared/widget-plugin-dropdown-filter/src/controllers/StaticComboboxController.ts
@@ -1,8 +1,8 @@
 import { DerivedPropsGate } from "@mendix/widget-plugin-mobx-kit/props-gate";
-import { StaticBaseController, StaticBaseControllerProps } from "./StaticBaseController";
+import { EnumBaseController, StaticBaseControllerProps } from "./EnumBaseController";
 import { ComboboxControllerMixin } from "./mixins/ComboboxControllerMixin";
 
-export class StaticComboboxController extends ComboboxControllerMixin(StaticBaseController) {
+export class StaticComboboxController extends ComboboxControllerMixin(EnumBaseController) {
     constructor({ gate }: { gate: DerivedPropsGate<StaticBaseControllerProps> }) {
         super({ gate, multiselect: false });
         this.inputPlaceholder = gate.props.placeholder ?? "Search";
